Rename routes to url & fix bugs (#45)
* fix auth, rename routing files per module to urls.py

* remove foreign keys from migration
