+++ src/auth/urls.py
@@ -1,4 +1,5 @@
 import base64
+import logging
 import os
 import urllib.parse
 
@@ -9,6 +10,8 @@ from auth import crud
 from constants import root_ip_address
 from fastapi import APIRouter, BackgroundTasks, HTTPException, Request
 from fastapi.responses import JSONResponse, RedirectResponse
+
+_logger = logging.getLogger(__name__)
 
 # ----------------------- #
 # utils
@@ -42,11 +45,12 @@ async def login_user(
     # verify the ticket is valid
     url = (
         f"https://cas.sfu.ca/cas/serviceValidate?service={urllib.parse.quote(root_ip_address)}"
-        f"/auth/login%3Fnext%3D{urllib.parse.quote(next_url)}ticket={ticket}"
+        f"/auth/login%3Fnext_url%3D{urllib.parse.quote(next_url)}&ticket={ticket}"
     )
     cas_response = xmltodict.parse(requests.get(url).text)
 
     if "cas:authenticationFailure" in cas_response["cas:serviceResponse"]:
+        _logger.info(f"User failed to login, with response {cas_response}")
         raise HTTPException(status_code=400, detail="authentication error, ticket likely invalid")
 
     else:
